Load add IParser types dynamically - no need to maintain a list.
